+++ CHANGELOG.md
@@ -9,6 +9,7 @@ There are no significant changes to which email addresses are considered valid/i
 * The dnspython package is no longer required if DNS checks are not used, although it will install automatically.
 * NoNameservers and NXDOMAIN DNS errors are now handled differently: NoNameservers no longer fails validation, and NXDOMAIN now skips checking for an A/AAAA fallback and goes straight to failing validation.
 * Some syntax error messages have changed because they are now checked explicitly rather than as a part of other checks.
+* The quoted-string local part syntax (e.g. multiple @-signs, spaces, etc. if surrounded by quotes) is now parsed but not considered valid by default. Better error messages are now given for quoted-string syntax since it can be confusing for a technically valid address to be rejected, and a new allow_quoted_local option is added to allow these addresses if you really need them.
 * Some other error messages have changed to not repeat the email address in the error message.
 * The library has been reorganized internally into smaller modules.
 * The tests have been reorganized and expanded. Deliverability tests now mostly use captured DNS responses so they can be run off-line.